+++ modules/terraform-aci-device-selection-policy/main.tf
@@ -17,6 +17,7 @@ resource "aci_rest_managed" "vnsRsLDevCtxToLDev" {
 }
 
 resource "aci_rest_managed" "vnsLIfCtx_consumer" {
+  count      = var.consumer_logical_interface != "" ? 1 : 0
   dn         = "${aci_rest_managed.vnsLDevCtx.dn}/lIfCtx-c-consumer"
   class_name = "vnsLIfCtx"
   content = {
@@ -27,8 +28,8 @@ resource "aci_rest_managed" "vnsLIfCtx_consumer" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToSvcRedirectPol_consumer" {
-  count      = var.consumer_redirect_policy != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_consumer.dn}/rsLIfCtxToSvcRedirectPol"
+  count      = var.consumer_logical_interface != "" && var.consumer_redirect_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_consumer[0].dn}/rsLIfCtxToSvcRedirectPol"
   class_name = "vnsRsLIfCtxToSvcRedirectPol"
   content = {
     tDn = "uni/tn-${var.consumer_redirect_policy_tenant != "" ? var.consumer_redirect_policy_tenant : var.tenant}/svcCont/svcRedirectPol-${var.consumer_redirect_policy}"
@@ -36,8 +37,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToSvcRedirectPol_consumer" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToBD_consumer" {
-  count      = var.consumer_bridge_domain != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_consumer.dn}/rsLIfCtxToBD"
+  count      = var.consumer_logical_interface != "" && var.consumer_bridge_domain != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_consumer[0].dn}/rsLIfCtxToBD"
   class_name = "vnsRsLIfCtxToBD"
   content = {
     tDn = "uni/tn-${var.consumer_bridge_domain_tenant != "" ? var.consumer_bridge_domain_tenant : var.tenant}/BD-${var.consumer_bridge_domain}"
@@ -45,8 +46,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToBD_consumer" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToInstP_consumer" {
-  count      = var.consumer_external_endpoint_group != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_consumer.dn}/rsLIfCtxToInstP"
+  count      = var.consumer_logical_interface != "" && var.consumer_external_endpoint_group != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_consumer[0].dn}/rsLIfCtxToInstP"
   class_name = "vnsRsLIfCtxToInstP"
   content = {
     redistribute = join(",", concat(var.consumer_external_endpoint_group_redistribute_bgp == true ? ["bgp"] : [], var.consumer_external_endpoint_group_redistribute_connected == true ? ["connected"] : [], var.consumer_external_endpoint_group_redistribute_ospf == true ? ["ospf"] : [], var.consumer_external_endpoint_group_redistribute_static == true ? ["static"] : []))
@@ -55,7 +56,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToInstP_consumer" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToLIf_consumer" {
-  dn         = "${aci_rest_managed.vnsLIfCtx_consumer.dn}/rsLIfCtxToLIf"
+  count      = var.consumer_logical_interface != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_consumer[0].dn}/rsLIfCtxToLIf"
   class_name = "vnsRsLIfCtxToLIf"
   content = {
     tDn = "uni/tn-${var.sgt_device_tenant != "" ? var.sgt_device_tenant : var.tenant}/lDevVip-${var.sgt_device_name}/lIf-${var.consumer_logical_interface}"
@@ -63,8 +65,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToLIf_consumer" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToSvcEPgPol_consumer" {
-  count      = var.consumer_service_epg_policy != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_consumer.dn}/rsLIfCtxToSvcEPgPol"
+  count      = var.consumer_logical_interface != "" && var.consumer_service_epg_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_consumer[0].dn}/rsLIfCtxToSvcEPgPol"
   class_name = "vnsRsLIfCtxToSvcEPgPol"
   content = {
     tDn = "uni/tn-${var.consumer_service_epg_policy_tenant != "" ? var.consumer_service_epg_policy_tenant : var.tenant}/svcCont/svcEPgPol-${var.consumer_service_epg_policy}"
@@ -72,8 +74,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToSvcEPgPol_consumer" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToCustQosPol_consumer" {
-  count      = var.consumer_custom_qos_policy != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_consumer.dn}/rsLIfCtxToCustQosPol"
+  count      = var.consumer_logical_interface != "" && var.consumer_custom_qos_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_consumer[0].dn}/rsLIfCtxToCustQosPol"
   class_name = "vnsRsLIfCtxToCustQosPol"
   content = {
     tnQosCustomPolName = var.consumer_custom_qos_policy
@@ -81,6 +83,7 @@ resource "aci_rest_managed" "vnsRsLIfCtxToCustQosPol_consumer" {
 }
 
 resource "aci_rest_managed" "vnsLIfCtx_provider" {
+  count      = var.provider_logical_interface != "" ? 1 : 0
   dn         = "${aci_rest_managed.vnsLDevCtx.dn}/lIfCtx-c-provider"
   class_name = "vnsLIfCtx"
   content = {
@@ -91,8 +94,8 @@ resource "aci_rest_managed" "vnsLIfCtx_provider" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToSvcRedirectPol_provider" {
-  count      = var.provider_redirect_policy != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_provider.dn}/rsLIfCtxToSvcRedirectPol"
+  count      = var.provider_logical_interface != "" && var.provider_redirect_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_provider[0].dn}/rsLIfCtxToSvcRedirectPol"
   class_name = "vnsRsLIfCtxToSvcRedirectPol"
   content = {
     tDn = "uni/tn-${var.provider_redirect_policy_tenant != "" ? var.provider_redirect_policy_tenant : var.tenant}/svcCont/svcRedirectPol-${var.provider_redirect_policy}"
@@ -100,8 +103,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToSvcRedirectPol_provider" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToBD_provider" {
-  count      = var.provider_bridge_domain != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_provider.dn}/rsLIfCtxToBD"
+  count      = var.provider_logical_interface != "" && var.provider_bridge_domain != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_provider[0].dn}/rsLIfCtxToBD"
   class_name = "vnsRsLIfCtxToBD"
   content = {
     tDn = "uni/tn-${var.provider_bridge_domain_tenant != "" ? var.provider_bridge_domain_tenant : var.tenant}/BD-${var.provider_bridge_domain}"
@@ -109,8 +112,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToBD_provider" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToInstP_provider" {
-  count      = var.provider_external_endpoint_group != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_provider.dn}/rsLIfCtxToInstP"
+  count      = var.provider_logical_interface != "" && var.provider_external_endpoint_group != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_provider[0].dn}/rsLIfCtxToInstP"
   class_name = "vnsRsLIfCtxToInstP"
   content = {
     redistribute = join(",", concat(var.provider_external_endpoint_group_redistribute_bgp == true ? ["bgp"] : [], var.provider_external_endpoint_group_redistribute_connected == true ? ["connected"] : [], var.provider_external_endpoint_group_redistribute_ospf == true ? ["ospf"] : [], var.provider_external_endpoint_group_redistribute_static == true ? ["static"] : []))
@@ -119,7 +122,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToInstP_provider" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToLIf_provider" {
-  dn         = "${aci_rest_managed.vnsLIfCtx_provider.dn}/rsLIfCtxToLIf"
+  count      = var.provider_logical_interface != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_provider[0].dn}/rsLIfCtxToLIf"
   class_name = "vnsRsLIfCtxToLIf"
   content = {
     tDn = "uni/tn-${var.sgt_device_tenant != "" ? var.sgt_device_tenant : var.tenant}/lDevVip-${var.sgt_device_name}/lIf-${var.provider_logical_interface}"
@@ -127,8 +131,8 @@ resource "aci_rest_managed" "vnsRsLIfCtxToLIf_provider" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToSvcEPgPol_provider" {
-  count      = var.provider_service_epg_policy != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_provider.dn}/rsLIfCtxToSvcEPgPol"
+  count      = var.provider_logical_interface != "" && var.provider_service_epg_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_provider[0].dn}/rsLIfCtxToSvcEPgPol"
   class_name = "vnsRsLIfCtxToSvcEPgPol"
   content = {
     tDn = "uni/tn-${var.provider_service_epg_policy_tenant != "" ? var.provider_service_epg_policy_tenant : var.tenant}/svcCont/svcEPgPol-${var.provider_service_epg_policy}"
@@ -136,10 +140,48 @@ resource "aci_rest_managed" "vnsRsLIfCtxToSvcEPgPol_provider" {
 }
 
 resource "aci_rest_managed" "vnsRsLIfCtxToCustQosPol_provider" {
-  count      = var.provider_custom_qos_policy != "" ? 1 : 0
-  dn         = "${aci_rest_managed.vnsLIfCtx_provider.dn}/rsLIfCtxToCustQosPol"
+  count      = var.provider_logical_interface != "" && var.provider_custom_qos_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_provider[0].dn}/rsLIfCtxToCustQosPol"
   class_name = "vnsRsLIfCtxToCustQosPol"
   content = {
     tnQosCustomPolName = var.provider_custom_qos_policy
+  }
+}
+
+resource "aci_rest_managed" "vnsLIfCtx_copy" {
+  count      = var.copy_logical_interface != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLDevCtx.dn}/lIfCtx-c-copy"
+  class_name = "vnsLIfCtx"
+  content = {
+    connNameOrLbl = "copy",
+    l3Dest        = var.copy_l3_destination == true ? "yes" : "no"
+    permitLog     = var.copy_permit_logging == true ? "yes" : "no"
+  }
+}
+
+resource "aci_rest_managed" "vnsRsLIfCtxToLIf_copy" {
+  count      = var.copy_logical_interface != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_copy[0].dn}/rsLIfCtxToLIf"
+  class_name = "vnsRsLIfCtxToLIf"
+  content = {
+    tDn = "uni/tn-${var.sgt_device_tenant != "" ? var.sgt_device_tenant : var.tenant}/lDevVip-${var.sgt_device_name}/lIf-${var.copy_logical_interface}"
+  }
+}
+
+resource "aci_rest_managed" "vnsRsLIfCtxToSvcEPgPol_copy" {
+  count      = var.copy_logical_interface != "" && var.copy_service_epg_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_copy[0].dn}/rsLIfCtxToSvcEPgPol"
+  class_name = "vnsRsLIfCtxToSvcEPgPol"
+  content = {
+    tDn = "uni/tn-${var.copy_service_epg_policy_tenant != "" ? var.copy_service_epg_policy_tenant : var.tenant}/svcCont/svcEPgPol-${var.copy_service_epg_policy}"
+  }
+}
+
+resource "aci_rest_managed" "vnsRsLIfCtxToCustQosPol_copy" {
+  count      = var.copy_logical_interface != "" && var.copy_custom_qos_policy != "" ? 1 : 0
+  dn         = "${aci_rest_managed.vnsLIfCtx_copy[0].dn}/rsLIfCtxToCustQosPol"
+  class_name = "vnsRsLIfCtxToCustQosPol"
+  content = {
+    tnQosCustomPolName = var.copy_custom_qos_policy
   }
 }