+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 - Add support for system performance settings
 - Add support for OSPF timer policies
 - Add support for OSPF route summarization
+- Add support for copy services (service graphs)
+- Add support for direct connect in service graph templates
 - Add support for OSPF area control
 
 ## 0.7.0